added new modal helper module to open and close add and edit forms. Updated vehicles add/edit modals to use the new helper methods
