+++ src/components/vehicle/list.js
@@ -3,6 +3,7 @@ import VehicleCard from "./vehicleCard"
 import VehicleManager from "../../api/dataManager";
 import "../../styles/vehicles/list.css"
 import ModalWrapper from "../modal/modalWrapper"
+import { modal } from "../../modules/modal/helpers"
 
 const VehiclesList = props => {
 
@@ -35,20 +36,12 @@ const VehiclesList = props => {
 
     const foundVehicle = vehicles.filter(filteredVehicle => filteredVehicle.id === vehicle.id);
 
+    document.querySelector(".modal--container").classList.add("show");
     document.querySelector(".modal-box").classList.add("show");
     document.querySelector(".modal-bg").classList.add("show");
 
     setFilteredVehicle(foundVehicle[0]);
   }
-
-  // Triggers add modal opening
-  const handleShow = () => {
-    setCreationView(true)
-
-    document.querySelector(".modal--container").classList.add("show");
-    document.querySelector(".modal-box").classList.add("show");
-    document.querySelector(".modal-bg").classList.add("show");
-  };
 
   // this reflects the vehicle update in the search list realtime by re-searching for the
   // vehicle after edit/delete
@@ -97,7 +90,7 @@ const VehiclesList = props => {
             </div>
           ) : null}
 
-            <button onClick={() => handleShow()} className="addDealership--btn">
+            <button onClick={() => modal.handleShow(setCreationView)} className="addDealership--btn">
                 Add New Vehicle
             </button>
         </div>
