+++ src/components/vehicle/modalEditForm.js
@@ -8,6 +8,7 @@ import FormControlLabel from '@material-ui/core/FormControlLabel';
 import FormControl from '@material-ui/core/FormControl';
 import SuccessSnackbar from "../modal/snackbar"
 import { errorHandler, validateForm} from "../validation/formValidator"
+import { modal } from "../../modules/modal/helpers"
 
 const VehicleEditModal = props => {
 
@@ -101,31 +102,10 @@ const VehicleEditModal = props => {
       VehicleManager.deleteUserData("vehicles", props.vehicle.id)
         .then(() => {
           props.setVehicleDeleted(true);
-          handleModalClose();
+          modal.handleEditClose(setEditMode);
         });
     }
   }
-
-  const handleModalClose = () => {
-    setEditMode(false);
-
-    const inputs = document.querySelectorAll('input')
-    const selects = document.querySelectorAll('select')
-    inputs.forEach(input => input.value = "")
-    selects.forEach(select => select.value = "none")
-
-    document.querySelector(".modal-box").classList.remove("show");
-    
-    setTimeout(() => {
-      document.querySelector(".modal-bg").classList.remove("show");
-    }, 300);
-
-    const muiSwitch = document.querySelector('.MuiSwitch-switchBase');
-
-    if (muiSwitch.classList.contains('Mui-checked')) {
-      muiSwitch.click();
-    }
-  };
 
   useEffect(() => {
     VehicleManager.getOne("vehicles", props.vehicle.id)
@@ -217,7 +197,7 @@ const VehicleEditModal = props => {
               <button 
                 className={`closeBtn ${props.vehicleEdited === true ? "disabled" : ""}`} 
                 disabled={props.vehicleEdited === true ? true : false}
-                onClick={handleModalClose}>
+                onClick={() => modal.handleEditClose(setEditMode)}>
                   Close  
               </button>
           </div>
@@ -261,7 +241,7 @@ const VehicleEditModal = props => {
                 <button onClick={handleSubmit} className="updateEmployee--btn">
                     Update
                 </button>
-                <button className="closeBtn" onClick={handleModalClose}>
+                <button className="closeBtn" onClick={() => modal.handleEditClose(setEditMode)}>
                     Cancel  
                 </button>
             </div>
